E-commerce-Backend-Development-codes/task2/requirements.txt
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Flask==3.0.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
pycparser==2.22
razorpay==1.4.2
requests==2.32.3
urllib3==2.2.1
Werkzeug==3.0.3
zipp==3.19.2
